refactor: ♻️ TrackRecord
